pyproject.toml
[tool.poetry]
name = "stock_analysis"
version = "0.1.0"
description = "股票分析"
authors = ["shabbywu <shabbywu@tencent.com>"]

[tool.poetry.dependencies]
python = "3.8.1"
futu-api = "^3.17.0"
pydantic = "^1.4"
influxdb = "^5.2.3"
SQLAlchemy = "^1.3.13"
aenum = "^2.2.3"
click = "^7.0"
jqdatasdk = "^1.8.1"
arrow = "^0.16.0"
isort = {extras = ["pyproject"], version = "^5.5.2"}
jinja2 = "^2.11.2"
backtrader = "^1.9.76"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
stock_analysis = 'stock_analysis.cli:cli'
[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
